+++ lib/jazz_hands/railtie.rb
@@ -4,33 +4,15 @@ require 'pry-git'
 require 'pry-remote'
 require 'awesome_print'
 
-if RUBY_VERSION >= '1.9.3'
-  require 'pry-nav'   # Enabled by default on MRI 1.9.3 only
+# Enable pry-nav by default on MRI 1.9.3 only
+require 'pry-nav' if RUBY_VERSION >= '1.9.3'
 
-  require 'coolline'
-  require 'coderay'
-end
 
 module JazzHands
   class Railtie < Rails::Railtie
     initializer 'jazz_hands.initialize' do |app|
       silence_warnings do
         ::IRB = Pry    # Replace IRB with Pry completely
-
-        # Use coolline with CodeRay for syntax highlighting as you type
-        # Only works on >= 1.9.3 because coolline depends on io/console
-        if RUBY_VERSION >= '1.9.3'
-          Pry.config.input = Coolline.new do |c|
-            c.transform_proc = proc do
-              CodeRay.scan(c.line, :ruby).term
-            end
-
-            c.completion_proc = proc do
-              word = c.completed_word
-              Object.constants.map(&:to_s).select { |w| w.start_with? word }
-            end
-          end
-        end
 
         # Use awesome_print for output, but keep pry's pager
         Pry.config.print = ->(output, value) do